% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/provincepoint.R
\name{provincepoint}
\alias{provincepoint}
\title{Mark points on Chinese province maps}
\usage{
provincepoint(cat, name, lat, lon, region = "anhui",
  theme = "sandsignika", title = "Example Map",
  titleAlign = "center", titleSize = "20px", titleColor = "#333333",
  subtitle = "", subtitleAlign = "center", subtitleSize = "",
  subtitleColor = "#666666", itermName = "Random data",
  headerFormat = "<small>{point.key}</small><table>",
  pointFormat = "<tr><td>Catagory: </td><td>: {series.name}</td></tr><tr><td>Location: </td><td>({point.lon:.2f}, {point.lat:.2f})</td></tr>",
  footerFormat = "</table>", hoverColor = "#a4edba",
  legendLayout = "horizontal", legendAlign = "center",
  legendTitle = "", legendVerticalAlign = "bottom", markerRadius = 3,
  mapNavigation = 1, mapNavigationVerticalAlign = "bottom",
  mapNavigationAlign = "left", width = NULL, height = NULL,
  elementId = NULL)
}
\arguments{
\item{cat}{category vector;}

\item{name}{name vector;}

\item{lat}{latitude vector;}

\item{lon}{longitude vector;}

\item{region}{A pinyi character, "anhui", "aomen", "beijing", "chongqing",
"fujian", "gansu", "guangdong", "guangxi", "guizhou", "hainan", "hebei",
"heilongjiang", "henan", "hubei", "hunan", "jiangsu", "jiangxi", "jilin",
"liaoning", "neimenggu", "ningxia", "qinghai", "shandong", "shanghai",
"shanxi", "shanxi2", "sichuan", "tianjin", "xianggang", "xinjiang",
"xizang", "yunnan", "zhejiang".}

\item{theme}{Chart theme, you can choose one from:
darkgreen/darkblue/avocado/darkunica/gray/
gridlight/grid/sandsignika/sunset;}

\item{title}{Chart title;}

\item{titleAlign}{The horizontal position of the title, such as "center";}

\item{titleSize}{The size of the title, such as "20px";}

\item{titleColor}{The color of the title, such as "#3333";}

\item{subtitle}{Subtitle of chart;}

\item{subtitleAlign}{The horizontal position of subtitles, such as "center";}

\item{subtitleSize}{The size of the subtitle, such as "16px";}

\item{subtitleColor}{The color of the subtitle, such as "#666666";}

\item{itermName}{Data attributes in tooltip;}

\item{headerFormat}{Header format, default value is '<small>{point.key}</small><table>';}

\item{pointFormat}{Point format, default value is "<tr><td>Catagory: </td><td>: {series.name}
</td></tr><tr><td>Location: </td><td>({point.lon:.2f}, {point.lat:.2f})</td></tr>";}

\item{footerFormat}{Footer format, default value is '</table>';}

\item{hoverColor}{The color of the area when the mouse is hovering.}

\item{legendLayout}{Legend, horizontal or vertical;}

\item{legendAlign}{Horizontal position of legend, center/left/right;}

\item{legendTitle}{The title of the legend;}

\item{legendVerticalAlign}{The vertical position of legends, top/center/bottom;}

\item{markerRadius}{Marker Radius;}

\item{mapNavigation}{show map navigation or not, 1 means show, 0 means not;}

\item{mapNavigationVerticalAlign}{Map navigation vertical align, top/bottom;}

\item{mapNavigationAlign}{map navigation align, center/left/right;}

\item{width}{Chart width;}

\item{height}{Chart height;}

\item{elementId}{NULL.}
}
\description{
provincepoint function provides a simple way to mark points on Chinese province map.
}
\examples{
library(hchinamap)
library(dplyr)
library(magrittr)
dir <- tempdir()
download.file('https://czxb.github.io/br/chinamappointdf.rda', file.path(dir, 'chinamappointdf.rda'))
load(file.path(dir, 'chinamappointdf.rda'), verbose = TRUE)
df <- chinamappointdf \%>\% dplyr::filter(province == "Shandong")
if(interactive()) {
  provincepoint(cat = df$cat, name = df$name,
      lat = df$lat, lon = df$lon, region = "shandong",
      title = "Urban distribution in Shandong Province",
      subtitle = "Data source: People's daily")
}

}
