% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hchinamap.R
\name{hchinamap}
\alias{hchinamap}
\title{hchinamap：Mapping China and its provinces, municipalities and autonomous regions using R and 'Highmaps'}
\usage{
hchinamap(name, value, region = NULL, width = NULL, height = NULL,
  elementId = NULL, itermName = "Random data", title = "",
  titleAlign = "center", titleSize = "20px", titleColor = "#333333",
  subtitle = "", subtitleAlign = "center", subtitleSize = "",
  subtitleColor = "#666666", min = 0, minColor = "rgb(255,255,255)",
  maxColor = "#006cee", legendLayout = "horizontal",
  legendAlign = "center", legendTitle = "",
  legendVerticalAlign = "bottom", hoverColor = "#a4edba",
  theme = "sunset")
}
\arguments{
\item{name}{Name vector of provinces or prefecture-level cities in China. The name of a province is short for two or three words. The name of a city does not include "市", such as "广州". For the municipality directly under the Central Government and the districts under the administrative region, the name should contain "区", such as "上海市黄浦区". You can find the corresponding name in the data set "chinadf" provided by the package.}

\item{value}{Value vector;}

\item{region}{You can choose "中国" "山西", "陕西", "内蒙古" and so on. There is no provincial data for Taiwan.}

\item{width}{Chart width;}

\item{height}{Chart height;}

\item{elementId}{NULL}

\item{itermName}{Data attributes in tooltip;}

\item{title}{Chart title;}

\item{titleAlign}{The horizontal position of the title, such as "center";}

\item{titleSize}{The size of the title, such as "20px";}

\item{titleColor}{The color of the title, such as "#3333";}

\item{subtitle}{Subtitle of chart;}

\item{subtitleAlign}{The horizontal position of subtitles, such as "center";}

\item{subtitleSize}{The size of the subtitle, such as "16px";}

\item{subtitleColor}{The color of the subtitle, such as "#666666";}

\item{min}{The minimum value of legend, 0 by default.}

\item{minColor}{The color corresponding to the minimum of the legend, such as "white";}

\item{maxColor}{The color corresponding to the maximum value of the legend, such as "#006cee";}

\item{legendLayout}{Legend, horizontal or vertical;}

\item{legendAlign}{Horizontal position of legend, center/left/right;}

\item{legendTitle}{The title of the legend;}

\item{legendVerticalAlign}{The vertical position of legends, top/center/bottom;}

\item{hoverColor}{The color of the area when the mouse is hovering.}

\item{theme}{Chart theme, you can choose one from:
darkgreen/darkblue/avocado/darkunica/gray/
gridlight/grid/sandsignika/sunset;}
}
\description{
hchinamap：Mapping China and its provinces, municipalities and autonomous regions using R and 'Highmaps'
}
\note{
Because the map data of Taiwan Bay have not been collated yet, it is impossible to draw provincial map of Taiwan Province for the time being.
}
\examples{
library(dplyr)
library(hchinamap)
library(magrittr)
china <- chinadf \%>\%
  dplyr::filter(region == "中国")
if(interactive()){
  hchinamap(name = china$name, value = china$value,
             title = "中国地图",
             minColor = "#f1eef6",
             maxColor = "#980043",
             subtitle = "czxa.top")
}
}
